+++ src/lib/Pym/PymChild.stories.svelte
@@ -1,0 +1,26 @@
+<script context="module">
+  import PymChild from "./PymChild.svelte";
+
+  export const meta = {
+    title: "Components/PymChild",
+    description: "A component that adds PymChild support to a page",
+    component: PymChild,
+    parameters: {
+      docs: {
+        description: {
+          component: "A basic headline component."
+        }
+      }
+    }
+  };
+</script>
+
+<script>
+  import { Story, Template } from "@storybook/addon-svelte-csf";
+</script>
+
+<Template let:args>
+  <PymChild {...args} />
+</Template>
+
+<Story name="Primary" />