+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Next
 
 - Simplify CSS in Toggle component
+- Add contribution guidance to Intro docs and README
+- Add Headline component
+- Add PymChild component
 
 ## v0.7.2
 
@@ -13,7 +16,6 @@
 - Changes snake_case variables to camelCase
 - Change BasicDropdown's value prop to accept null
 - Add Navbar case for no projectUrl with story documentation
-- Add Headline component
 
 ## v0.7.1